Add a no side effect support for Ray
